Refactor

+++ app.rb
@@ -13,7 +13,6 @@ TS_FPS           = 24
 HLS_SEGMENT_TIME = 2
 
 $channel_id = nil
-$ffmpeg_thread = nil
 $ffmpeg_pid = 0
 $rectest_pid = 0
 
@@ -26,8 +25,10 @@ def stop_ffmpeg
 end
 
 def start_rectest
-  puts "start_rectest ch = #{$channel_id}"
+  return if $rectest_pid > 0
+
   Thread.start do 
+    puts "start rectest (ch = #{$channel_id})"
     if $channel_id.nil?
       Open3.popen3(RECTEST_PATH, '/udp', '/udpport', RECTEST_PORT.to_s) do |i, o, e, w|
         $rectest_pid = w.pid
@@ -42,37 +43,37 @@ def start_rectest
 end
 
 def start_ffmpeg
-  puts "start_ffmpeg"
-  $ffmpeg_thread = Thread.new do
-    loop do
-      Dir.glob("#{WWW_PATH}/*.ts").each do |f|
-        File.delete f
-      end
-      Open3.popen3(FFMPEG_PATH,
-                   '-i', "udp://127.0.0.1:#{RECTEST_PORT}?pkt_size=262144^&fifo_size=1000000^&overrun_nonfatal=1",
-                   '-f', 'mpegts',
-                   '-threads', 'auto',
-                   '-map', '0:0', '-map', '0:1',
-                   '-acodec', 'libvo_aacenc', '-ar', '44100', '-ab', '128k', '-ac', '2',
-                   '-vcodec', 'libx264', '-s', '1280x720', '-aspect', '16:9', '-vb', '2m',
-                   '-r', TS_FPS.to_s,
-                   '-g', "#{TS_FPS * HLS_SEGMENT_TIME}",
-                   '-force_key_frames', "expr:(t/#{HLS_SEGMENT_TIME})",
-                   '-f', 'segment',
-                   '-segment_format', 'mpegts',
-                   '-segment_time', HLS_SEGMENT_TIME.to_s,
-                   '-segment_list', "#{WWW_PATH}/#{M3U8_FILENAME}",
-                   '-segment_list_flags', 'live',
-                   '-segment_wrap', '50',
-                   '-segment_list_size', '5',
-                   '-break_non_keyframes', '1',
-                   "#{WWW_PATH}/stream%d.ts") do |i, o, e, w|
-                     puts "ffmpeg is running (pid = #{w.pid})"
-                     $ffmpeg_pid = w.pid
-                   end
-      puts "ffmpeg is dead"
-      $ffmpeg_pid = 0
+  return if $ffmpeg_pid > 0
+
+  Thread.start do
+    puts "start ffmpeg"
+    Dir.glob("#{WWW_PATH}/*.ts").each do |f|
+      File.delete f
     end
+    Open3.popen3(FFMPEG_PATH,
+                 '-i', "udp://127.0.0.1:#{RECTEST_PORT}?pkt_size=262144^&fifo_size=1000000^&overrun_nonfatal=1",
+                 '-f', 'mpegts',
+                 '-threads', 'auto',
+                 '-map', '0:0', '-map', '0:1',
+                 '-acodec', 'libvo_aacenc', '-ar', '44100', '-ab', '128k', '-ac', '2',
+                 '-vcodec', 'libx264', '-s', '1280x720', '-aspect', '16:9', '-vb', '2m',
+                 '-r', TS_FPS.to_s,
+                 '-g', "#{TS_FPS * HLS_SEGMENT_TIME}",
+                 '-force_key_frames', "expr:(t/#{HLS_SEGMENT_TIME})",
+                 '-f', 'segment',
+                 '-segment_format', 'mpegts',
+                 '-segment_time', HLS_SEGMENT_TIME.to_s,
+                 '-segment_list', "#{WWW_PATH}/#{M3U8_FILENAME}",
+                 '-segment_list_flags', 'live',
+                 '-segment_wrap', '50',
+                 '-segment_list_size', '5',
+                 '-break_non_keyframes', '1',
+                 "#{WWW_PATH}/stream%d.ts") do |i, o, e, w|
+                   puts "ffmpeg is running (pid = #{w.pid})"
+                   $ffmpeg_pid = w.pid
+                 end
+    puts "ffmpeg is dead"
+    $ffmpeg_pid = 0
   end
 end
 
@@ -85,17 +86,21 @@ def delete_all
 end
 
 def restart_all
-  puts "restart_all"
   stop_rectest
   stop_ffmpeg
   start_rectest
-  start_ffmpeg if $ffmpeg_thread.nil?
+  start_ffmpeg
 end
 
 ### 
 
 delete_all
 restart_all
+
+Signal.trap("INT") do |signo|
+  stop_ffmpeg
+  stop_rectest
+end
 
 get '/' do
   File.read(File.join('public', 'index.html'))
